Add name field to user.
This is the @username for the user. We require that it be at least 3 characters long and no more than 20, that it contain only alphanumerics and _, and that it be case-insensitively unique amongst all users. This last bit requires some extra work because ActiveRecord doesn't support the creation of case-insensitive db indices.

See http://stackoverflow.com/questions/7948501/case-insensitive-unique-index-in-rails-activerecord for details. And use rake db:structure:dump and db:structure:load to move the database structure around.