Make "code" the primary key for invitations.
See http://stackoverflow.com/questions/1200568/using-rails-how-can-i-set-my-primary-key-to-not-be-an-integer-typed-column

The previous scheme tried have a normal integer id for the table, overriding the find and to_param methods. But in test I found that #reload failed (it called find with some additional params), and this seemed like a slippery slope. The above required hacking the migration, but it's the real deal: code is now the primary key field for the table, which ActiveRecord respects through and through.