Rename team_memberships to players.
This is the beginning of a larger change to turn the Player model into a first-class entity, not just an association model. The Player has an optional user associated with it, reflecting the reality that not all players known to the system have been claimed by users of the system.

We still need to work out the invitation and claim mechanism, and rip out the "fabricated user" mechanism in the Player model.