Deal with one-to-many embedded models.
EmbeddedModel extends Backbone.model. It allows subclasses to specify a models() function, which returns a hash. The keys in the hash specify names of attributes that come from the server (in JSON) that represent embedded models. The values identify either a Backbone model class or an instance of a Backbone collection.

If the value is a Backbone model and the instance doesn't yet exist in the parent, it is instantiated with the given parameters. If the instance does already exist, we call set() on it with the given parameters.

If the value is a Backbone collection, we call set() on the collection with the given list of object parameters.