Fix bug in handling of goal/penalty deletions.
Here's what was happening before:

When the model (either goal or penalty) noticed that it was being removed from a collection, it would destroy itself. The view would notice the model being destroyed, and remove itself.

That's wrong. The problem is that calling destroy on the model causes a sync (DELETE) on the server. But really we only want whoever is initiating the delete to invoke the server method. What was happening was that every observer would receive the updated list of goals/penalties, see that a goal was missing and so remove it from the collection, which would trigger a delete/destroy. Bah.

So, the way it works now:

 * clicking the delete button in the UI causes the model to be destroyed. That invokes the server DELETE method which eventually uses the async messaging to update the list of goals all around.
 * whenever someone receives an updated list of goals, we generate remove events on models; when the view sees its model being removed from a collection, it removes itself from the view.
 * Presumably, the model, which is now no longer in a view nor in a collection, gets garbage collected. But we don't call destroy on it, because that would attempt to do something on the server, which is not desired.