Fix player model to generate invitations.
Previously, we generated an invitation on the creation of a new player only if the creator provided an email address that didn't correspond to an existing user in the system. That is, if the creator identified a user, we simply attached that user to the player and continued.

Now, we'll generate an invitation whenever a player is created if the creator provides either an @username, or an email address.