Implement system_names instead of user hack.
The system_names table now maintains a unique index of all @names in the system. In addition to the name itself, for each record it also stores a polymorphic link back to the object named. Currently, these names point to users and teams, but they will also likely point to locations.

The big win here, beyond a cleaner architecture, is that we can render feeds with @name links to users, teams, etc., without a database lookup. Instead of de-referencing the name to determine what url to generate, we generate a link in the form of /system_names/#{name}. The system_names_controller does the lookup and redirect as necessary, meaning that the db lookup is essentially deferred until the user clicks on the link, a vast improvement over the previous situation.