Officials can belong to exactly one league.
I'm not sure what I was thinking about when I implemented this earlier, but an official isn't a user -- it's just a name. So there's no reason for an official to belong to more than one league. In this respect, an official is analogous to a player. We'll probably want to allow a user to claim an official, so a single user who officiates for more than one league can be two officials. But we don't really care the two representations match the same user.