Improve updating of penalty views.
Two things here.

First, render views when the tab is shown, which is both more efficient and forces tab panes to be updated to accurately reflect current state when the user opens the tab.

Second, re-render the currently visible pane whenever the game restarts. This causes penalties which have recently expired to disappear from the current view.