Default to showing current penalties.
This is a tough one: what's really most interesting to the marker is the current penalties, as these represent players who are in the penalty box. So defaulting to displaying it makes sense.

The downside is that when the marker creates a non-timed penalty (an ejection, for instance), he doesn't see any immediate confirmation that the penalty was created. Perhaps a solution to this problem is to display the number of penalties in the tab.