Have client sync timer expirations back to server.
The current implementation using Delayed::Jobs to manage timer expiry is not sufficiently accurate. It seems like the worker wakes up every few seconds to look for jobs, which means that there's typically a delay of several seconds after the expiration before the system notices that a timer has expired.

To remedy, we have the client-side timer post an AJAX sync request back to the game controller, which it forwards to the game. The game in turn talks to each timer (for the game clock plus each running penalty) to give it the opportunity to see whether it has expired.

(We *could* be more targeted, setting up a timer controller that the timer talks to specifically to see whether it has expired, but it's easier to manage the authorization of the action through the game: the marker for the game is the only one allowed to make the request.)

For now, I'm leaving in place the Delay::Job mechanism to allow it to sweep up anything that isn't covered by the existing mechanism; this new mechanism only works if there's an active marker marking the game when the expiration occurs.