Immediately re-render penalties view on model chg.
I was operating under the misunderstanding that players in the penalty box can't re-enter the ice until there's a stoppage of play. Since that's incorrect, we're simply going to re-render the current view whenever any of the underlying model changes; as a result, when a penalty expires, it will disappear from the Current view and appear in the Expired view immediately.