Calculate and store "advantage" for each goal.
"Advantage" is an integer associated with each goal. It refers to the number of players the team that scored the goal has on the ice minus the number of players on the ice for the opposing team. (It's actually calculated by looking at the number of held off the ice by minor penalties.)

So, if an advantage of +1 refers to a power play, 0 means even-handed, -1 means short-handed.