Prevent premature submission of marker forms.
I was kind of surprised to see that my disabled save button in the penalty editor was actually still making a round-trip server request -- the request failed validation so there was no visible effect for the user. But it turns out that adding the disabled class to a link affects only the the appearance. So I had to tell the backbone views to trigger the save action only when :not(.disabled) via css.