Batch broadcasts of game state changes.
The core implementation is simple: call game.batch_broadcasts, passing it a block of code that might cause one or more broadcasts of state changes. batch_broadcasts works with game.broadcast_changes to ensure all of the requests for broadcasts are collected together into a single final broadcast that includes all the state that changed.

For this to work, however, all the broadcasting has to go through the same instance of game, and this is where active record's "inverse_of" feaure is supposed to help us but fails. It seems that inverse_of doesn't work as soon as you add a scope to the mix. So we have to jump through some hoops to prevent multiple instances of the same game object from getting into the chain.

It does seem to work, though. Scoring a goal, which before could result in a total of 5 broadcasts (2 during phase I, where the clock is stopped and the goal is recorded, 3 during phase II, where the goal is updated, a penalty is canceled, and an activity feed item is created), now slims down to one broadcast in each phase.

More importantly, starting the clock, which could result in a multiple broadcasts as penalties are (re-)started now also happens in a single broadcast, which should help to keep penalty timers better synced.