Allow timers to slave to a master.
Basically, we want the penalty timers to be slaved to the game clock. That means that the penalty timers record as their start time the same time that game clock does, and the same goes for pausing.

The goal of all of this is two-fold: one, to allow the marker to control actual stop-times across the board, and to ensure that on start all the timers start at the same instant.

This entails a few changes. First, when the marker sends up its record of elapsed_time when the game is stopped, we use that to calculate the wall-clock time (the time that was on the wall clock at the moment that the marker pressed the button) and we store that calculated time in paused_at. Previously, we were fudging the seconds_paused value to account for any differences.

Second, the penalty timers look first to the game clock when they are setting paused_at and calculating seconds_paused. In this way, multiple timers running simultaneously should be both more accurate and perfectly in sync with each other.