Rework goal editor for efficient data entry.
The basic idea is to use the player chiclets that we developed for the penalty editor. But since we now have 3 full sets of player choices, we wrap them in what amounts to an accordion display so that only one set is displayed at a time.

Further, we allow the user to clear or edit any item. When this yields an invalid set of input, we disable the save button to prevent the user from continuing.