Rejigger User#name.
The field that was once called User#name is now cached_system_name, which is intended to be a local copy of the corresponding name field from the attached SystemName record.

There's now a now field (which is also called name) that corresponds to the user's (optional) real name. When reading this attribute, we default to the user's system name if no name is provided.