We can now play and pause…
AdobeEdge.getComposition("EDGE-404308").getStage().play("Loop") and .stop()

I ripped the sound out because Eric had it loading a new sound library as part of an event handler on the timeline, which can't be the right way to do it. And I couldn't get the sound to stop when I stopped the animation. So… we'll just manage audio ourselves.