Revamp activity feed machinery.
There are now a collection of subclasses of ActivityFeedItem, found in models/feed, that are implemented via STI.

Each class must implement the following methods:

message: returns a text string that is the content of the feed's message
mentioned_objects: called during creation, it returns an array of objects mentioned in this item

The array of mentioned objects is turned at create time into an array of mentions that point at each mentioned object, and thus can be used for indexing purposes (as it works today).

In the next step, we remove the references to mentioned objects and replace with their @names, stored directly in the item. This will allow us to render an item without making any further database calls.