Change store_location to a before_filter.
The problem is that if we call authenticate_user! we don't wind up storing the location of the page that the user was trying to navigate to. By storing the location in a before_filter, the user winds up going to the page they tried to navigate to, not the page they navigated from.